% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-sib2.R
\name{call_sib2}
\alias{call_sib2}
\title{Run SiB2}
\usage{
call_sib2(file, out.dir, id.sim, run.pars, ini.pars, soil.pars, morpho.pars,
  physio.pars, veg.pars, verbose = TRUE)
}
\arguments{
\item{file}{A character string with the Input file with meteorological forcing dataset}

\item{out.dir}{A character string with the path where the output file ('sib2diag<id>.txt') is saved.}

\item{id.sim}{A character string to identify simulation (e.g. '_run1').}

\item{run.pars}{A numeric vector with model setup parameters
* `date1`: 1960111601
* `date2` : 1961123024
* `lai.max` : 65
* `aeromet` : 1
* `isnow` : 0
* `ipbl` : 1
* `ilw` : 3
* `itrunk` : 20
* `ivtype` : 6
* `istype` : 3
* `idirr` : 0}

\item{ini.pars}{initial conditions}

\item{soil.pars}{soil parameters}

\item{morpho.pars}{morphological parameters}

\item{physio.pars}{physiological parameters}

\item{veg.pars}{vegetation parameters (miscelaneous)}

\item{verbose}{logical, default is TRUE, to print messages}
}
\value{
A string character with the output file name ('sib2diag<id>.txt')
}
\description{
Run SiB2
}
\examples{
input <- system.file("extdata", "data2", package = "sib2r")

}
