% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib2r.R
\docType{package}
\name{sib2r}
\alias{sib2r}
\alias{sib2r-package}
\title{sib2r: A R interface to the Simple Biosphere Model 2.0 (SiB2).}
\description{
A multi-OS and stand-alone version of the SiB2 model.
}
\section{inmetr functions}{

The only function you're likely to need from \pkg{sib2r} is
\itemize{
   \item \code{\link{call_sib2}}  
}
 The data included in \code{sib2r} is available at:
\itemize{
   \item \file{sib2r/inst/extdata/data2}  
}
}

